Refactoring Span/ReadOnlySpan and Extension methods
